% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{lf_check_api}
\alias{lf_check_api}
\title{Is an API visible for a given application?}
\usage{
lf_check_api(app_name, api_name)
}
\arguments{
\item{app_name}{Application name.}

\item{api_name}{API name.}
}
\value{
\code{TRUE} or \code{FALSE}. Same cache-delay caveat as
  \code{lf_check_app()}.
}
\description{
Is an API visible for a given application?
}
