% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{lf_check_main_thread}
\alias{lf_check_main_thread}
\title{Is the simulated main thread running?}
\usage{
lf_check_main_thread()
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Is the simulated main thread running?
}
