% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{lf_data_create}
\alias{lf_data_create}
\title{Create a data handle}
\usage{
lf_data_create(api_name)
}
\arguments{
\item{api_name}{API name the handle is bound to.}
}
\value{
An external pointer handle.
}
\description{
Create a data handle
}
