% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{lf_data_free}
\alias{lf_data_free}
\title{Release a data handle}
\usage{
lf_data_free(hnd)
}
\arguments{
\item{hnd}{Data handle to release.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Release a data handle
}
