% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{lf_data_get_size}
\alias{lf_data_get_size}
\title{Total size of a data handle}
\usage{
lf_data_get_size(hnd)
}
\arguments{
\item{hnd}{Data handle.}
}
\value{
The buffer size in bytes, as an integer.
}
\description{
Total size of a data handle
}
