% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{lf_data_read}
\alias{lf_data_read}
\title{Read up to n bytes from a data handle}
\usage{
lf_data_read(hnd, n)
}
\arguments{
\item{hnd}{Data handle.}

\item{n}{Maximum number of bytes to read.}
}
\value{
A raw vector of length at most \code{n}.
}
\description{
Read up to n bytes from a data handle
}
