% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{lf_data_write}
\alias{lf_data_write}
\title{Append bytes to a data handle}
\usage{
lf_data_write(hnd, data)
}
\arguments{
\item{hnd}{Data handle.}

\item{data}{Raw vector to append.}
}
\value{
The number of bytes written, as an integer.
}
\description{
Append bytes to a data handle
}
