% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{lf_is_loaded}
\alias{lf_is_loaded}
\title{Is the LingoFuse runtime loaded?}
\usage{
lf_is_loaded()
}
\value{
\code{TRUE} if the runtime has been loaded, \code{FALSE}
  otherwise.
}
\description{
Is the LingoFuse runtime loaded?
}
