% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{lf_last_error}
\alias{lf_last_error}
\title{Last error reported by the runtime}
\usage{
lf_last_error()
}
\value{
A length-1 character vector. The string may be empty if no
  error has been recorded.
}
\description{
Last error reported by the runtime
}
