% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{lf_notify_bin}
\alias{lf_notify_bin}
\title{One-way notification (binary payload)}
\usage{
lf_notify_bin(app_name, api_name, req_raw)
}
\arguments{
\item{app_name}{Target application name.}

\item{api_name}{Target API name.}

\item{req_raw}{Request payload as a raw vector.}
}
\value{
\code{NULL}, invisibly.
}
\description{
One-way notification (binary payload)
}
