% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{lf_prepare_service}
\alias{lf_prepare_service}
\title{Prepare a listening service}
\usage{
lf_prepare_service(listen_addr, physics_addr = listen_addr)
}
\arguments{
\item{listen_addr}{Local binding address.}

\item{physics_addr}{Address advertised to clients. Defaults to
\code{listen_addr}.}
}
\value{
The internal tag assigned to the service, invisibly.
}
\description{
Prepare a listening service
}
